Bump min PHP version and include lockfile
The new version requirement will allow newer versions of dependencies to be used, which will increase compatibility with other products.

The lockfile ensures that the same deps will be installed everywhere, be it on another dev's machine, or on the CI server. This way we can actually test with the deps that are going to be shipped, instead of a random set.