Mention enabling vaulting from the PayPal account only when it was disabled there
We were always displaying the text 'To use vaulting features, you must enable vaulting on your account' next to the checkbox for enabling vaulting. This message is irrelevant if vaulting is already enabled in the account. This commit makes that message to be displayed only when vaulting is disabled in the account.